requirements.txt
torch
transformers[torch]
hydra-core
omegaconf
datasets
wandb
pytest
